+++ polybench-code/linear-algebra/kernels/atax/atax.c
@@ -18,7 +18,7 @@ void init_array (int m, int n,
 		 DATA_TYPE POLYBENCH_2D(A,M,N,m,n),
 		 DATA_TYPE POLYBENCH_1D(x,N,n))
 {
-  int i, j;
+  long long i, j;
   DATA_TYPE fn;
   fn = (DATA_TYPE)n;
 
@@ -37,7 +37,7 @@ void print_array(int n,
 		 DATA_TYPE POLYBENCH_1D(y,N,n))
 
 {
-  int i;
+  long long i;
 
   POLYBENCH_DUMP_START;
   POLYBENCH_DUMP_BEGIN("y");
@@ -59,7 +59,7 @@ void kernel_atax(int m, int n,
 		 DATA_TYPE POLYBENCH_1D(y,N,n),
 		 DATA_TYPE POLYBENCH_1D(tmp,M,m))
 {
-  int i, j;
+  long long i, j;
 
 #pragma scop
   for (i = 0; i < _PB_N; i++)
