+++ polybench-code/linear-algebra/solvers/trisolv/trisolv.c
@@ -19,7 +19,7 @@ void init_array(int n,
 		DATA_TYPE POLYBENCH_1D(x,N,n),
 		DATA_TYPE POLYBENCH_1D(b,N,n))
 {
-  int i, j;
+  long long i, j;
 
   for (i = 0; i < n; i++)
     {
@@ -38,7 +38,7 @@ void print_array(int n,
 		 DATA_TYPE POLYBENCH_1D(x,N,n))
 
 {
-  int i;
+  long long i;
 
   POLYBENCH_DUMP_START;
   POLYBENCH_DUMP_BEGIN("x");
@@ -59,7 +59,7 @@ void kernel_trisolv(int n,
 		    DATA_TYPE POLYBENCH_1D(x,N,n),
 		    DATA_TYPE POLYBENCH_1D(b,N,n))
 {
-  int i, j;
+  long long i, j;
 
 #pragma scop
   for (i = 0; i < _PB_N; i++)
