+++ polybench-code/stencils/adi/adi.c
@@ -17,7 +17,7 @@ static
 void init_array (int n,
 		 DATA_TYPE POLYBENCH_2D(u,N,N,n,n))
 {
-  int i, j;
+  long long i, j;
 
   for (i = 0; i < n; i++)
     for (j = 0; j < n; j++)
@@ -34,7 +34,7 @@ void print_array(int n,
 		 DATA_TYPE POLYBENCH_2D(u,N,N,n,n))
 
 {
-  int i, j;
+  long long i, j;
 
   POLYBENCH_DUMP_START;
   POLYBENCH_DUMP_BEGIN("u");
