+++ polybench-code/stencils/jacobi-2d/jacobi-2d.c
@@ -18,7 +18,7 @@ void init_array (int n,
 		 DATA_TYPE POLYBENCH_2D(A,N,N,n,n),
 		 DATA_TYPE POLYBENCH_2D(B,N,N,n,n))
 {
-  int i, j;
+  long long i, j;
 
   for (i = 0; i < n; i++)
     for (j = 0; j < n; j++)
@@ -36,7 +36,7 @@ void print_array(int n,
 		 DATA_TYPE POLYBENCH_2D(A,N,N,n,n))
 
 {
-  int i, j;
+  long long i, j;
 
   POLYBENCH_DUMP_START;
   POLYBENCH_DUMP_BEGIN("A");
